[k1c] Optimize build_vector for v4i16 and v4f16
Summary:
A vector immediate is composed of all the vector member immediates. Afterwards,
non constant values are inserted in the correct position, using INSF or OR
instruction (for optimization purpose, if first or last position are registers
then OR instruction is used to materialise the constant, elsewhere INSF is
used).

Test Plan: CI

Reviewers: O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D2966