[kvx] Fix link flags for compiler-rt tests
Summary:
The mathematic library needs to be linked twice since some tests are
referencing symbols of the libm, that reference symbols which have to
be defined by compiler-rt.

For example, to create an executable from test.c, the compiler will
link it as `test.o compiler-rt.a -lm`, which fails if test.c reference
a libm symbol that reference a symbol of compiler-rt.a. The solution
here is to add another -lm to satisfy the libm definition of test.o
before linking with compiler-rt.a, such like: `test.o -lm
compiler-rt.a -lm`.

Test Plan: CI

Reviewers: O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D8598