[kvx][tca] Add tca builtin types
Summary:
Defines three new architecture dependent
types for handling tca types. Their names
and generated IR types are:
	- __tca256   ->   <256  x i1>
	- __tca512   ->   <512  x i1>
	- __tca1024  ->   <1024 x i1>

TCA Types for 128 and 64 bits types are not
visible for the user, as they can not be
manipulated easily or or saved to the stack.
TCA types also can't be passed as arguments,
only being passed as pointers.

Test Plan: added tests

Reviewers: O81 llvm-coolidge, lthevenoux

Reviewed By: O81 llvm-coolidge, lthevenoux

Subscribers: lthevenoux

Differential Revision: https://phab.kalray.eu/D9012