[kvx] Allow unaligned memory access
Summary:
This patch fixes the KVXTargetLowering as to allow unaligned memory
accesses. They are always considered as fast for any alignment >= 256
bits.
It also changes the target structure alignment, setting alignment to
256bits to elements of size of 512 and 1024 bits.
These are memory alignments, as in-registers, the alignment is defined
by the associated registers to data types.

This improves multiples useless stack allocation when it was trying
to generate aligned accesses.

Test Plan: added testes

Reviewers: lthevenoux, O81 llvm-coolidge

Reviewed By: lthevenoux, O81 llvm-coolidge

Subscribers: vray, aodinot, nbrunie

Differential Revision: https://phab.kalray.eu/D9224