[kvx][tca] Add TCA builtins
Summary:
This patch extends Clang's builtin parser to be able to handle
target specific types. Now, a type started with T passes the
intrinsic types string to a parser to the TargetInfo, expecting
the type and the string pointer in the correct position.

This adds all requested TCA intrinsics, except lvc and swap. Swap
will require some extra thinking in how to restrain the copy
operations to be preserved inside a bundle, as the usual lowering
happend before the packing of bundles.

As for the builtins that perform inplace changes to a vectors, not
returning a type seems the most logical manner, else it could give
the user a false sensation that he could preserve the previous value.
Perhaps, if required, that can be implemented, performing a vector
copy before, but I think that was not the expeted behavior.
The builtins that do not return are:
	moveto[hi|lo]
	convdhv[0-1]
	convwbv[0-3]
	fmma242hw[0-3]

We also add operations that perform all the sub-parts operations,
and returns a new TCA vector/wide type, them being:
	moveoto
	moveto
	convdhv
	convwbv
	fmma242

Test Plan: added tests

Reviewers: lthevenoux, O81 llvm-coolidge

Reviewed By: lthevenoux, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D9300