[kvx][TCA] Add swapvfwo builtin / intrinsic
Summary:
The swpvfwo instrinsic performs a swap of a QuadReg and a VectorReg.
It requires three instructions to be executed in a single bundle.
To achieve this, from the intrinsic, we pattern match it to a
pseudo-instruction, that is expanded just before the final instruction
selection, packed in a bundle instruction.
The packetizer it self did not expect to see pre-define bundles, and
required working on. Now it will close the existing bundle.

Test Plan: added tests

Reviewers: O81 llvm-coolidge, lthevenoux

Reviewed By: O81 llvm-coolidge, lthevenoux

Maniphest Tasks: T14222

Differential Revision: https://phab.kalray.eu/D9441