[kvx][TCA] Ref T15038: fix swap TCA and Core regs
Summary:
swapvfwo was taking arguments as reference. However, in C that does
not mean the compiler assumes the value did change. Future uses
of the values end up using the previous value.

This patch changes the intrinsic to obtain the pointer to the TCA
and Core values to be swaped, using load and store of the values,
hopping that mem2reg will elimitate these simple patters.

Test Plan: existing tests

Reviewers: lthevenoux, O81 llvm-coolidge!

Subscribers: mqho

Differential Revision: https://phab.kalray.eu/D9655