[kvx] Add patterns for ADDX 2,4,8,16 HQ,WP RR,RI
Summary:
Detect both RR and RI variants of all vector ADDX instructions.
Whenever the immediate does not fit in a KVX 32 bit splat vector, it will
write the immediate to a register and use addx instead of (add(sll)) or
(or(sll)).

It also introduces tests to detect immediate splat_ and splat_32
vectors. As well vectors of immediate values that fit in
1,2,3 or 4 bits.
This is required as some of the `add` instructions are turned
to `or` so it needs to detect patterns such as:
  (add (shl))
as well as
  (or (shl), (insideLimitImmediateValuesVector))

Test Plan: added tests / existing tests

Reviewers: O81 llvm-coolidge, lthevenoux

Reviewed By: O81 llvm-coolidge, lthevenoux

Subscribers: lthevenoux

Differential Revision: https://phab.kalray.eu/D10529