[kvx][FIX] Fix instructions that MaybeCall
Summary:
This adds further rules to detect instructions that may turn into a call.
The inline asm was removed as it is captured as a function call already.
It also refactors such that we can decide if a given intrinsic callsite may
not lower to a function call based in the calling instruction attributes.
For last a small refactoring such that when hwloop is disable, no analysis
is done recursively in the loops to decide if they can be converted to hw loops.

Test Plan: existing tests

Reviewers: O81 llvm-coolidge, lthevenoux

Reviewed By: O81 llvm-coolidge, lthevenoux

Subscribers: lthevenoux

Differential Revision: https://phab.kalray.eu/D10802