[kvx] Refactor KVX pseudo instructions expansion
Summary:
Merge the pseudo instruction expansion passes in a single one,
controled by a KVX::STAGE argument.
Calling it at addPreSched2 it is ensured that loads/stores are expand
before the scheduler. It does make a marginal change in Coremak,
from 2625.44 to 2626.05, so not much of a big change.
We also need to add a third pass of kvx pseuso instruction expansion
for ENDLOOP, which needs to be done after the cfg optimizations.

Test Plan: tests changed, CI

Reviewers: lthevenoux, O81 llvm-coolidge

Reviewed By: lthevenoux, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D11674