[kvx] Make LSR care first about code size
Summary:
This tells to the LSR that minimizing the number of instructions is
more important than minimizing the register pressure, as KVX has a quite
large number of registers.
It also moves Immediate and Scale costs to the least important factors
when comparing two solution candidates.

Test Plan: existing tests, ci, benchmarks

Reviewers: O81 llvm-coolidge, lthevenoux

Reviewed By: O81 llvm-coolidge, lthevenoux

Subscribers: lthevenoux

Differential Revision: https://phab.kalray.eu/D10819