[kvx] Add patterns for ADDUWD ADDWD
Summary:
Adds patterns for ADDUWD ADDWD instructions, except for immediate
variants.

It also defines table-gen methods for discovering if a value can be
consided as an extended value, checking if we know if at least half
of the upper bits are either zero or one or it is a zero/sign extended
operation.

It also marks multiple add operations with the isAdd flag in table-gen.

Implements: T16404

Test Plan: added/changed tests

Reviewers: lthevenoux, O81 llvm-coolidge

Reviewed By: lthevenoux, O81 llvm-coolidge

Maniphest Tasks: T16404

Differential Revision: https://phab.kalray.eu/D12818