[kvx] Improve/add vector element/subvector extract patterns
Summary:
Adding patterns for vector element extract when the index
is not a constant value.

Vector extract elements and subsextract elements define that
remaining padding bits have `undefined` value. So extracting the
element 0 is actually a NOP, but returning the input register would
resolve in type mismatch when lowering. So a `COPY` instruction is
used, as these will be optimizied out further on the pipeline.

For the remaining elements simple `SRLW` or `SRLD` do replace
`EXTFZ`, as these are `TINY` instead of `LITE` operations.

Implements: T16404

Test Plan: changed tests

Reviewers: lthevenoux, O81 llvm-coolidge

Reviewed By: lthevenoux, O81 llvm-coolidge

Subscribers: jvillette

Maniphest Tasks: T16404

Differential Revision: https://phab.kalray.eu/D13014