[kvx] Use llvm infrastructure to generate ffma
Summary:
InstCombine and DagCombine do query the targetLowering to decide if
fusing multiply-add is beneficial for a given data type.
This allows us to better control when we do not want to fuse, such as when
multiplying and adding a `float4` vector, where we have both add and mul
operations, but not a fma.
This also adds patterns for ffmahq for types `half4`, `half2` and `half`.
Also cleared custom-combine for `fadd` and `fsub` that existed soely to
generate `FMS` as now we can detect it on patterns.

Test Plan: added tests

Reviewers: O81 llvm-coolidge, bddinechin

Reviewed By: O81 llvm-coolidge, bddinechin

Subscribers: lthevenoux, bddinechin, nbrunie

Differential Revision: https://phab.kalray.eu/D10954