[kvx] Add reg. indexed extract vector from vNi8 vNi16
Summary:
This adds patterns to perform vector element extraction
indexed by a value in a register to vectors with elements
`i8` and `i16`. Different then the other types, which are
legal, these are promoted to `i32` values, so they must be
correctly zero extended to clear up the upper bits.

Implements: T16803

Test Plan: changes tests

Reviewers: lthevenoux, csix, O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D14123