[kvx] Set boolean formats
Summary:
This defines the boolean contents for scalar boolean results to be `0` or `1`.
It also defines the vector boolean results for `kvx-v2` to be `0` or `-1`.
This allows to correctly clear the padding bits when reading and input
argument extended from a `i1` or returned by a function.
However `f16` uses `v4f16` comparison instructions, so the result must be
cleared before use. So it adds patterns to avoid duplicated zeroing of the
upper bits.

It also avoid `sign-extending` vector results.

Test Plan: changed tests

Reviewers: O81 llvm-coolidge, lthevenoux

Reviewed By: O81 llvm-coolidge, lthevenoux

Subscribers: lthevenoux

Differential Revision: https://phab.kalray.eu/D14296