[kvx] Allow using `fabs/fneg` to perform `xor/and`
Summary:
This defines the TargetLowering hook `hasBitPreservingFPLogic` enabling
the usage of `fabs` instead of `and` and `fneg` instead of `xor` when
one of the arguments is a huge integer that performs the exact same
thing. This is enabled for scalar values of bit sizes of `16, 32 & 64`.

Implements: T16803

Test Plan: added tests

Reviewers: csix, lthevenoux, O81 llvm-coolidge

Reviewed By: lthevenoux, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D14334