[kvx] Add `TargetLowering::shouldReplaceBy`
Summary:
This adds the hook `TargetLowering::shouldReplaceBy`
which can be used in a generic manner to decide if
a given input `SDNode` should be replaced by a node
of `ToOpcode` of the same `ValueType`. It can be used
to guide instructions expansions, such as when a custom
implementation is defined for `ROTL` and `ROTR`, where
both of them would actually expand, entering in a infinite
loop, as well can be used to prevent `DAGCombiner` to
replace Nodes that are more benefitial to the target, such
as vector multiply by a vector shift.

To keep remaining targets behaviour, the default function
always returns `true`.

Test Plan: changed tests

Reviewers: lthevenoux, O81 llvm-coolidge!

Subscribers: csix

Differential Revision: https://phab.kalray.eu/D14568