[kvx][NFC] Reorder conditional store/load operands, add memory instruction attributes
Summary:
This patch sets the last two arguments of all conditional memory
operations (loads/stores/atomics) being the test condition and the
test value.

This allows to easily convert a non-predicated instruction into a
predicated on, by simply swapping the machine instruction opcode
and appending the tested condition and predicate at the end.

It also adds attributes `mayLoad` to all our load operations,
`mayStore` to all our store operations and both of them to
the atomic operations, allowing to easily decide when we
are dealing with a memory operations.
It also adds attribute `predicable` to memory
operations that can be transformed into a predicated one,
so we can easily decide when we can if-predicate them.

Implements: T16998

Test Plan: n/a

Reviewers: lthevenoux, csix, O81 llvm-coolidge

Reviewed By: lthevenoux, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D14660