[kvx] Add `v4[if]32` register extract element
Summary:
This adds patterns to allow extracting a vector element defined by
a value in register. First it decides if we obtain the first or second
half of the vector with a `CMOVED`. Second it shifts the required
amount of bit required to put the first or second element at the lower
part of the register.

It also prefers to use `srld` instead of `srad` for extracting elements.

Implements: T16803

Test Plan: changed tests

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D14717