[kvx] Add patterns to match `FMULWDC`
Summary:
This adds four patterns to match `FMULWDC`, two of them to match the
`real` part and other two to match the `imaginary` part. The reason
behind it is that `v2f64` can't be passes in our calling convention, so
whenever just having this operation in a function, it would not generate a
single vector result, but 2 scalars.

Implements: T16803

Test Plan: changed tests

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D14998