[kvx][NFC] Prevent test from assembling file
Summary:
SemaCXX/warn-unused-local-typedef-serialize.cpp only checks semanitcs.
However it is wasting resources as it goes all the way to assembling the
file. This changes it to only perform syntax validation and quit.
It also prevents failure in kvx as it does not have an assembler in the
testing environment.

Test Plan: N/A

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D15458