[kvx][EH] Implement SjLj exception builtins lowering
Summary:
This implements the lowering the builtins generated by the
`SJLJ Exception Handling preparation` pass. It consist in:

- `builtin_sjlj_setjump`: Code that spills and reload all registers.

- `builtin_sjlj_longjump`: Code that actually performs the jump,
restoring `FP`, `SP` registers and jumps to the correct `PC (IP)`.

- `builtin_sjlj_setupDispatch`: Code that populates the jump table with
the function landing pads and generates code to obtains the landing
pad address and jump to it.

This patch also touches the register info to obtain a list of all registers
that must be spilled during an exception.

Implements: T14358

Test Plan: added tests

Reviewers: csix, O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D17696