[kvx] Add side effects to addcd and sbfcd intrinsics
Summary:
ADDCD and SBFCD dependencies on $CS gets ignored during loop-invariant code
motion. As such, addcd and sbfcd builtins might be wrongly moved out of loops
because they appear to be loop-invariant, when they actually are not.

This adds side effects to addcd and sbfcd intrinsics to prevent that. This
fixes the problem, at the cost of worse scheduling when addcd and sbfcd
builtins are involved.

A better solution involving chains in the DAG might be implemented later.

Test Plan: added frontend test (the wrong transformation happens during IR generation)

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Maniphest Tasks: T18408

Differential Revision: https://phab.kalray.eu/D18727