[kvx] Enable optimizations for splat vectors
Summary:
KVX architecture relies in the instruction to define the number
of bits being used from a 64bit register. That being so, every
element of the vector at `bit_start_position % 64 == 0` can
be extracted without any cost. Telling this to the compiler
allow it to improve operations with splat vectors, by
performing a single element operation and latter splatting the
result.

Test Plan: changed tests

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D19478