[kvx][loopdo] Add test case for empty hardware loops bug + Add fix for the bug
Summary:
This commit adds a test case where llc generates an empty hardware loop
This commit also proposes a fix for this bug.
The fix detects empty loops at IR level by counting instructions and
ignoring debug instructions, phi nodes, and pseudo instructions. If this
counted number is 3 (increment+compare+branch) then the loop is empty
and it is considered not profitable to convert it into a hardware loop,
thus allowing it be optimized away later in the pipeline.
Ref: T18770

Test Plan: To be compared with updated version after bug fix

Reviewers: csix, dsampaio, O81 llvm-coolidge

Reviewed By: csix, dsampaio, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D19515