[kvx] Adds cv2 Shift Right Symmetric instructions and cv[12] patterns
Summary:
This adds cv2 `srsbos` instructions and immediate patterns for all
types, which can be achieved at C language by divisions of a power
of 2 value (constant value).

It required to change division from promoting to expanding, which
requires the library to implement word divisions.

It also requires to implement the TargetLowering hook `BuildSDIVPow2`
which prevents the compiler to convert the power of 2 divisions into
a sequence of shift operations.

Test Plan: added/changes test[s], CI

Reviewers: O81 llvm-coolidge, csix

Reviewed By: O81 llvm-coolidge, csix

Subscribers: bddinechin

Maniphest Tasks: T18531

Differential Revision: https://phab.kalray.eu/D19643