[kvx] Fix itinerary latencies
Summary:
The itinerary operand latencies were slightly wrong. An ALU instruction would
be typically expressed as [3, 2, 2], modeling the write at E1, the read at RR.

But this is not how the schedulers interpret it. Instead, it is
interpreted as the write having a latency of 3, and the read happening 1 cycle
later. Meaning the effective latency of two successive and dependent ALU
instructions is 2 cycles.

This patch decrements all write latencies by 1 to fix this.

Test Plan: updated CodeGen tests

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D21103