[kvx] Add isSoloInstruction hook in TargetInstrInfo
Summary:
We add a isSoloInstruction hook which returns true if the instruction
cannot be issued at the same cycle as any other instruction. This hook
defaults to false.

We add support of this hook in ScoreboardHazardRecognizer to ensure any
SUnit whose MachineInstruction satisfies isSoloInstruction is not issued
at the same time as another instruction.

We implement this hook for the KVX architecture.

This support only works for instructions who have SUnits. Instructions
such as DEBUG_VALUE still needs to be moved manually out of a bundle.

Test Plan: updated CodeGen tests

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D22043