[kvx] Packetize instructions with respect to schedule cycle
Summary:
- Create KVXPostPacketizer, KVXPostScheduler and KVXScheduleDAGMI classes
  - KVXPostPacketizer contains logic for bundling based on cycle information
  - KVXPostScheduler records such cycle information
  - KVXScheduleDAGMI extracts CFI instructions to put them back after scheduling
    and activates ScheduleDAG mutations.
- Add KVXPostScheduler pass where KVXPacketizer was. Deactivate PostRA
  MachineScheduler (as it is now run with KVXPostScheduler)
- Add enterRegion and leaveRegion virtual methods in MachineScheduler. These
  communicate to the scheduling strategy that a region is entered or left.
- Add support of isSchedulingBoundaryPostRA in MachineScheduler
- Enforce prolog and epilog not to be scheduling boundaries in PostRA
- Add artificial edges to ensure prolog and epilog are respectively at the
  beginning/end of their region in PostRA
- Recover the data dependency edges from SUs to ExitMI

The gist of this patch lies in KVXPostScheduler class, a decorator over
PostGenericScheduler. As instructions are scheduled, the cycle information is
saved. When leaving a scheduling region, the bundles are created according to
the saved cycle information. During bundle creation, instructions that do not
belong in a bundle (such as debug instructions) are moved to the bottom of it.

ExitMI is not part of the schedule, so we must bundle it with the last bundle
with an ad-hoc approach: we recover the data dependency edges through a
ScheduleDAGMutation named ExitMIDataDeps, then we check these edges, scheduling
resources and a few edge cases to decide whether we bundle ExitMI. We also
check for stalls.

When CFI instructions are encountered, we temporarily remove them from
the basic block and then put them back. Indeed, CFI instructions are treated as
scheduling boundaries to prevent them from being moved above the instruction
performing the state change they describe. This was detrimental to the schedule,
as it was creating unnecessary regions; in the case of our backend, this affected in
particular the prolog (the GET instruction emission of our pseudo-instruction
expansion pass emits a GET and a CFI).

We add code in KVXInstrInfo::isSchedulingBoundary to avoid treating
the prolog and epilog as scheduling boundaries. Indeed, the former
KVXPacketizer was transgressing all scheduling boundaries. The scheduler is
stricter, so such modification is required to bundle the prolog and epilog.
Since we want to avoid the epilog moving too much (decrementing SP in the
middle of the function), we insert artificial edges with a new
ScheduleDAGMutation that we name PrologEpilogFreeze. This ensures the prolog
and epilog do not move without compromising the bundling.

This patch is preferable to the former KVXPacketizer who had a greedy approach
to bundling, which was detrimental to performance because of breaking
the schedule (including an instruction that makes the pipeline stall in a
packet that had no stalling instructions yet).

Ignoring CFI instructions when scheduling and removing prolog/epilog scheduling
boundaries result in better code as well.

Test Plan: update CodeGen tests, add cfi-cb.ll and icall.ll tests, weekly CI

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Subscribers: mgligor, mkapfer

Maniphest Tasks: T17357

Differential Revision: https://phab.kalray.eu/D21058