[kvx] Support dependent multi-stage reservation in ScoreboardHazardRecognizer
Summary:
ScoreboardHazardRecognizer did not support reserving two units of the
same kind in different stages of the itinerary. This was causing issues
in the KVX backend for GET instruction reserving 2 TINY resources.

We add such support by adding a
ScoreboardHazardRecognizer::shouldUpdateBetweenStages method: if true,
the hazard recognizer updates the reservation tables between each stage
of the instruction. By default, this support is turned off.

We create a KVXHazardRecognizer which has this support turned on.

Test Plan: no change

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D22042