[kvx] Fix a wrong emission of MAKEi43, constant too big
Summary:
The function isSplatBuildVec, used for checking that an integer is a valid
splat build-vector immediate; and in particular checking its size; was
incorrectly assuming the integer to be unsigned.

The MAKE instruction of build_vector assumes the integer is signed.
Hence, we had off-by-one errors in the size of constants given to MAKE,
resulting in too big bundles in unfortunate cases.

Test Plan: update jfctint.ll, all others are unchanged

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Subscribers: hkhallouki

Maniphest Tasks: T19709

Differential Revision: https://phab.kalray.eu/D22358