[kvx][NFC] Rename `SCALARCOND` and add set CPU
Summary:
`SCALARCOND` and `LSUCOND` do have the same arguments and
values, so the compiler can treat them equally. However, there are
many more instructions using `LSUCOND` other than `SCALARCOND`
(`CMOVED` only). So this renames the modifer to be consistent
with most instructions.

This also defines the default kvx `CPU` string to `kv3-1`, implements
the target method to check if the current cpu string is equals to a
given string and also implements a new target method to obtain the
current cpu string.

Test Plan: n/a

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Maniphest Tasks: T18267

Differential Revision: https://phab.kalray.eu/D23220