[kvx] Remove ACB4.8 compatibility, add KVX_BUILTIN
Summary:
This removes the ability of the compiler to warn and process
TCA instrinsics that have been deprecated in ACB4.8.

It introduces a new builtin declaration marco, `KVX_BUILTIN` which
generates the code to emit the instruction, handling cpu specific
builtins as well the modifier string validation automatically.

It restores `lv` and `sv` variant builtins for `kv3-1` cpu as these
are not 1:1 mapped to `xlo` and `xso`, including different modifiers.

It also renames all uses of `swapvo` to `xswap256`.

Test Plan: changed tests, CI

Reviewers: O81 llvm-coolidge, csix

Reviewed By: O81 llvm-coolidge, csix

Maniphest Tasks: T18267

Differential Revision: https://phab.kalray.eu/D23234