[kvx][cv2] Create 4k tca buffers
Summary:
This joins TCA and Core sub-reg indexes freeing 2 bits in the lane mask,
allowing the creation of SubRegIndexes of up to 32 sub-lanes. Now the
creation of larger registers would require expanding the number of possible
lanes from 64 to 128, changing the lanemask type. That should not be a
simple task.

Apart from the ability of creating larger TCAbuffer, this patch is a NFC.

Test Plan: nfc

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D24577