[kvx][builtins] Avoids crashing if not constant string
Summary:
It's accepted that we pass a nullpointer to the string modifier,
but we can't pass a pointer in register, it must be
a constant string. This avoids breaking the compiler in such cases.

This undoes D24809 `ffmaxdwd` -> `ffmadwd` as it was not a typo.

Maniphes Tasks: T20349

Test Plan: ci

Reviewers: csix, O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D24979