[kvx] Support of ready builtin
Summary:
Implements `__builtin_kvx_ready` support in frontend and backend.

This builtin returns a value whose computation involves all of its arguments.

For that, we add a new `ready` intrinsic (with vararg) and emit bitcasts in the
frontend to ensure arguments with legal types are emitted (e.g. bitcasting a
v8i32 to a v4i64).

In the backend, we custom lower the relevant `ready` intrinsic patterns into
one of the four possible `READYp` pseudo instructions (respectively with 1, 2,
3 or 4 arguments), each with scheduling information.

Finally, these instructions are expanded after PostRA MachineScheduler (at the
same time as BUNDLE expansions) into one or two `ORd` instructions.

It can be used in conjunction with (not yet implemented) store and storec
builtins to ensure the memory transactions involving the memory load of the
value are finished before executing the memory store.

Test Plan: added clang and llc CodeGen tests

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Subscribers: jhascoet

Maniphest Tasks: T20278

Differential Revision: https://phab.kalray.eu/D24416