[kvx] Support unconditional __builtin_kvx_store* variants
Summary:
Implements support for the unconditional store builtins.

There are two variable arguments (in addition to their positional arguments):
  - When only one v.a. is present, it may be either a volatile flag (if
    constant) or a dependency (e.g. from a `ready` builtin)
  - When two v.a. are present, the first is assumed to be a dependency, the
    second is a volatile flag.

We emit regular stores when we can. Under the presence of the ready argument,
we emit a new kvx.store intrinsic. That intrinsic takes the store value, a
pointer, the size of the store (8 for sb, 16 for sh, etc..) and a fake argument
for the "ready" dependency.

That intrinsic is then lowered to a pseudo-instruction STOREp (STOREpv if
volatile) with similar arguments. We use `variable_ops` in the
pseudo-instruction definition to encode the "ready": this ensures the fake
dependency has 0 latency, as all operations involving arguments in variable_ops
are treated as having 0 latency.

Finally, the pseudo-instruction STOREp (or STOREpv) is expanded into the
adequate store instruction, after post-RA scheduling.

This patch also lowers i128 stores into v2i64 store (by bitcasting) to ensure `sq` is selected.

Test Plan: added CodeGen tests

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Maniphest Tasks: T20335

Differential Revision: https://phab.kalray.eu/D24635