[kvx][cv2] Add all `ffdm*` instructions and intrinsics
Summary:
This adds all variants of `ffdm[as]w[pq]?` instructions
and intrinsics.

The macro `KVX_MANY_BUILTIN` generates polymorphic
intrinsics based on the type of the return value. So
this teaches tablegen to generate a argument type
with twice the number of elements of a given one,
or build a vector of 2 elements of a given scalar,
named `LLVMDoubleElementsOfVecOrScalar`.

Test Plan: added/changed tests, CI

Reviewers: O81 llvm-coolidge!, csix

Maniphest Tasks: T20349, T18267, T18467, T18475, T18464, T18471, T18483

Differential Revision: https://phab.kalray.eu/D25806