[kvx] Any little thing counts: any builtins
Summary:
Adds the and family of builtins that verify in integer vector if
it contains at least one element that is zero, or not zero.

Documents the KVX_MANY_BUILTIN in the BuiltinsKVX.def.

Expands KVX_MANY_BUILTIN with two abilities: Overload now
specifies if the intrinsic is overloaded or not, and if it is
overloaded, it defines from which result/argument comes the
type to be used in the overload. If the value/argument being
used is split, then the resulting type is used.

It also allows the Parts field to be a negative value, meaning
that the builtin arguments should be split into a single intrinsic
call, not altering the result value.

Test Plan: added tests, CI

Reviewers: csix, O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D25836