[kvx][builtins] XSplat that!
Summary:
This adds the tca vector/buffer splat of a single 64 bit value
family of builtins. It detects if the value is a immediate to
start populating the 256 bits with a cv2's `xsplatdo`. In CV1
it enforces reusing the 1024 bit matrix transpose to copy
matrix registers, instead of allowing the compiler to generate
COPY, which will be split into sub-registers copy.

Test Plan: added tests

Reviewers: O81 llvm-coolidge!, csix

Maniphest Tasks: T20349

Differential Revision: https://phab.kalray.eu/D25933