[kvx] Fix tablegen types
Summary:
These fixes in tablegen are NFC in llvm 12,
but they wouldn't compile in llvm 15.

- `KVX_INSTRUCTION -> Instruction`
Many of the instructions are not declared as KVX_INSTRUCTION a sub-class
of Instruction. This avoids invalid casts to the child class of Instruction
structures.

- `PatFrag -> PatFrags`
Same as above, PatFrag is a sub-class of PatFrags accepting a single
element list.

- `SRS`
Correctly uses register variant of instruction in place of immediate
one.

Test Plan: n/a

Reviewers: csix, O81 llvm-coolidge!

Maniphest Tasks: T20650

Differential Revision: https://phab.kalray.eu/D26746