[kvx] [ADD/SUB]x fixes and better mul/shl control
Summary:
`addx/subx` extending patterns as the shift/multiply occurs
after the value is extended. However, if the shift is marked
with nsw/nuw the original patterns are still correct.

It prevents converting `mul` to `shl` when it might generate
a single fused multiply add/sub or when it is a vector operand
and the multiplied values do not hold a splat value.

Test Plan: changed tests, CI

Reviewers: O81 llvm-coolidge, csix

Reviewed By: O81 llvm-coolidge, csix

Maniphest Tasks: T20650

Differential Revision: https://phab.kalray.eu/D26900