[kvx][FIX] Removes `SELECTp` and `Wrapped37W`
Summary:
The `SELECTp` pseudo-instruction patterns are wrong, as it accepts
both immediate and register operands to what, in the pseudo instruction,
it is just registers. From now the lowering goes directly to `CMOVEDrr`
instructions. Adding immediate variants should be done in a case-by-case
as the immediates should be sign-extended to 64 bits, as declared in
the instruction. This would break verify-machineinstrs every time the
pseudo would be used with an immediate.

It also replaces operand `Signed37W` by `Wrapped32` as the're identical.

Test Plan: ci

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Maniphest Tasks: T20650

Differential Revision: https://phab.kalray.eu/D27010