[kvx][FIX] Do not set value as undef
Summary:
When marking the load of a spilled value the resulting register was
being marked with `DefineNoRead`, which means, it is being defined
and it is `undefined` valued. This makes WideRegSpill.ll
break in llvm-15's machine instruction verification. This fixes it to not
mark the loaded value as `undef`.

Test Plan: ci

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D26984