[kvx] No more elf writter
Summary:
This removes kvx elf writter, as it is not possible to emit elf, as
instructions encodings in table-gen are missing. This avoids unsupported
tests. It is required to add unsupported in a few test files that assume
that the default target is capable to generate object files or that it
supports embedding clang ast into object files.

This makes all regression tests to pass or not run, allowing to remove
the xfail tests in integration.

Test Plan: ci

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Maniphest Tasks: T18156

Differential Revision: https://phab.kalray.eu/D27109