[kvx][ABI] Set MPPANative raw value to 1.
Function attribute `MPPANative` raw value was 1 << 63, which prevents
using of the function attribute NoFree. Using value 1, also the value of
a return attribute, their usage won't conflict, so a single raw value
represents Zext for a return value and MPPANative for a function.

Perhaps an improvement from this is to split raw attributes per subject
(argument / return ....) which would allow a finer validation of raw
attributes values, as well allowing much more values.

Reviewers: csix