[kvx] Introduce KVXCodeGenPrepare
Summary:
This adds a last IR transformation that prepares the code for KVX code
selection, performing instruction-combine alike pattern matching and
transformations.

The first use is the case of vector_compare + vector_reduction, using
vNi1 intermediary vectors, which are not alike the hardware, that
preserves the inputs size and are converted to bitcast of invalid size.
The transformation detects the pattern
`(icmp ( bitcast (icmp vector))`
and converts it to
`(icmp (or reduce (icmp vector))`.

Test Plan: added test

Reviewers: O81 llvm-coolidge!, csix

Maniphest Tasks: T20872

Differential Revision: https://phab.kalray.eu/D30263