[kvx] Enable MachinePipeliner
Summary:
This patch enables software pipelining in our LLVM backend.

It is not yet finely tuned: the focus of this patch is on functional
correctness and having no performance regression.

Changes:

- add a new variant of `analyzeBranch` and implement it on KVX.

MachinePipeliner relies on the hook `analyzeBranch` to analyze the loop branch;
to make it work on hardware loops, `analyzeBranch` needs to tackle the ENDLOOP
case. However, doing so opens a pandora box of CFG optimizations that we do not
want in hardware loops.

This change adds a boolean to distinguish between MachinePipeliner and other
passes.

- add new hook `getPipelinerMaxMII` and implement it on KVX to let target
  control how big a pipelined loop can be
- add ScalarcondMod and ComparisonMod enums in KVXInstrInfo.h to improve
  readability
- add MachinePipeliner pass right after KVXHardwareLoops
- add support for ENDLOOP in insertBranch() and removeBranch()
- add isReg() checks in getBaseAndOffsetPosition()
- implement getIncrementValue() hook
- implement analyzeLoopForPipelining() hook with a KVXPipelinerLoopInfo class
  to perform the transformations asked by software pipelining on a loop. These
  mostly involve modifying the trip count and generating condition on the trip
  count.
- add -enable-kvx-pipeliner and -disable-kvx-pipeliner options to respectively
  force the enabling or disabling of MachinePipeliner

Test Plan:
- updated CodeGen tests
- WEEKLY CI passed,
- added enable-pipeliner.ll to test the new options
- moved all tests related to passes in a new Passes folder
- good performance results (F2573008)

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Maniphest Tasks: T20310

Differential Revision: https://phab.kalray.eu/D27456