[kvx][upstream] Fix slp vectorizer bug
Summary:
The SLP vectorizer might recognize that smaller vector elements might be
used without precision loss in integer operations. It can happen that a
cast or extending instruction might now be a truncating operation.
When asking the backend the cost of a expanding instruction, the actual
result type was actually smaller thant the original type. This fixes the
opcode if required, avoid passing the instruction if changed.

Test Plan: added test, CI

Reviewers: csix, O81 llvm-coolidge!

Maniphest Tasks: T21742

Differential Revision: https://phab.kalray.eu/D30348