[kvx] Fix kvx loadc builtins lowering
Summary:
The custom lowering would generate a result node with load type
different than the original builtin call when `tryMatchLoadStoreValue`
changed the value of `Lv` to an operand of different value type.
This fixes this by always using the desired result of the matched
operation.

Test Plan: N/A

Reviewers: csix, O81 llvm-coolidge!

Maniphest Tasks: T21847

Differential Revision: https://phab.kalray.eu/D30939