[kvx] Fix requesting cost of invalid extensions
Summary:
This fixes two places where the vectorizers do request the cost of
invalid extending instructions. One is reported upstream, the other
is still to be understood the reason, but this produces a rather valid
roundabout by using cost when the extended and source types are
the sames or by updating the instruction opcode if required.

Test Plan: added tests, CI

Reviewers: csix, O81 llvm-coolidge!

Maniphest Tasks: T21742

Differential Revision: https://phab.kalray.eu/D30966