[kvx][llvm15][slp] Update scalar cost
Summary:
Following upstream fix 0a68cd23049de778a4440168c855aff38f202940,
we are only allowed to use the smaller type if all the operations can
be done on it, and the scalar cost should also be updated to the new
opcode. This produces the same tests changes from upstream. Should be
reverted when passing to llvm 18.

Test Plan: changed test

Reviewers: csix, O81 llvm-coolidge!

Maniphest Tasks: T21742

Differential Revision: https://phab.kalray.eu/D31001