[kvx][cv1] Make fminnum and fmaxnum legal
Summary:
CV1's `fmin` instructions do not fit in any OpenCL or IEEE standards
when handling `NaN`. The same for `fmax`. However it is possible
to use them with any of the 3 standards if it is known that there are
no `NaN` operands. This patch covers the OpenCL, implemented under the
selection dag nodes `fmaxnum` and `fminnum`. In such standard, whenever
a `NaN` is compared to a ordered value, the ordered value is returned.
And whenever two `NaN`, signaling or not, are the operands, it does not
matter which one is returned. As well `-0.0` and `0.0` ordering does not
matter. This makes such nodes by implementing 2 distinct patterns.
Whenever marked with `hasNoNaN`, the direct instruction is matched. The
second pattern perform two comparisons, which decides if a given operand
is either `NaN` or both operands are ordered and one is smaller than the
other. Always that a `NaN` is detected the value returned is the other.

For: `fmin`:
`Predicate1 = X is a NaN`
`Predicate2 = operands (X,Y) are ordered and X > Y`
`result = (Predicate1 | Predicate2) ? Y : X`

For: `fmax`:
`Predicate1 = Y is a NaN`
`Predicate2 = operands (X,Y) are ordered and X > Y`
`result = (Predicate1 | Predicate2) ? X : Y`

As for deciding if a value is `NaN` the value in question is shifted one
bit to the left to remove the sign bit and making the exponent start from
the upper bit. Following [1], [2], [3] the `NaN` have all exponent bits set
to `1` and has a non-zero fraction part, so a single unsigned comparison of
greater than is capable to decide if the floating point is inside a `NaN`.

[1] https://en.wikipedia.org/wiki/Half-precision_floating-point_format#Exponent_encoding
[2] https://en.wikipedia.org/wiki/Single-precision_floating-point_format#Exponent_encoding
[3] https://en.wikipedia.org/wiki/Double-precision_floating-point_format#Exponent_encoding

Test Plan: changed tests

Reviewers: csix, O81 llvm-coolidge!

Maniphest Tasks: T22038

Differential Revision: https://phab.kalray.eu/D32048