[kvx][CostModel] Multiple improvements
Summary:
Set `phi` nodes with `vNi1` results as invalid, as to prevent vector comparison
results being passed from one basic block to another, as their all need to be
truncated to an `vNi8` from their different sources of vector comparison.

Check if the index for (sub) vector (element) insert / extract is (unsigned)(-1),
if so assume it is variable and holds cost of 3. Also element inserts has at
least a cost of 1, as it is an `insf` or a `copyd` at least.

The same goes for extract elements with extends, except they have a minimal
cost of 1.

`fabs`, `fgetsign` also use binops under the hood and have a bitwidth of
256 bits per cycle.
`fadd` and `fsub` have bitwidth of 64/128 depending of being cv1 or cv2.
The same for intrinsics `fmuladd` and `fma`.

Set `icmp` of vectors of elements smaller than `i8` illegal.

Test Plan: added/changed tests, coremark -> 3416, amat_v1 -> 16596

Reviewers: O81 llvm-coolidge!, csix

Maniphest Tasks: T22038, T21688

Differential Revision: https://phab.kalray.eu/D32429