[kvx][llvm 15] Handle `ctpop(bitcast(vcmp))`
Summary:
At kvx-codegen-prepare convert such pattern to the pattern
`bitConvert(neg(add_reduce(sext(vcmp))))`. The `bitConvert`
is not required when there is a single user that expects
the type of the `neg` operation. See that `vcmp` generates
`-1 or 0` per lane, that is why the accumulation is negated.

At dag-combine the `sext` and `vcmp` are combined into a
single `SETCC` already producing the correct vector size.

DCE is executed at each changed basic block preventing
DAG-Combine to loop infinitely in dead code that won't be
replaced.

A few errors regarding `vcmp` result bits being `0 / 1` or
`0 / -1` are inserted and will be fixed in the next patch.

Test Plan: added / changed tests

Reviewers: csix, O81 llvm-coolidge!

Maniphest Tasks: T21688, T20650

Differential Revision: https://phab.kalray.eu/D32481