[kvx][CostModel] Bug fixes and workarounds
Summary:
- The SLPVectorizer holds a bug when a reduction
is generated but one of the values used inside
the reduction is also used outside. This bug is
mostly triggered when reducing predicate vectors
(from vector comparisons), binop combining them
and extending them to be used as masks.
Setting the cost to invalid for extensions of
bitmasks that are not directly generated from a
comparison instruction, or that do not extend
to the same element width of the compared
vectors.

- Divisions costs might be used when the cost
for intrinsics not directly handled by the target
TTI. In such cases the division arguments will be
missing so accessing `Args[1]` must be checked.

- This adds `fshl` and `fshr` intrinsic costs to
avoid passing to the decomposed version which
will compute the cost as scalars as the shifts
arguments will be missing.

- The SLPVectorizer still requests casts with invalid
arguments, such as truncate with source arguments
of same size or smaller than the produces size.
Instead of aborting the TTI returns an invalid cost.

Test Plan: added / changed tests

Reviewers: O81 llvm-coolidge!, csix

Maniphest Tasks: T21688, T20650

Differential Revision: https://phab.kalray.eu/D32623