[kvx] Fix bidirectional scheduling by introducing new hook
Summary:
Because of how it is implemented, right now, bidirectional scheduler does not
really schedule bidirectionally. It does take Bot and Top queues, but as soon
as Bot is picked once, Top will never get a chance as TopCand.Reason will
always be NoCand.

This patch fixes that by restoring the old value of TryCand.Reason if TryCand
was not picked.

This fix alone did not have much impact because Bot still has the priority over
Top unless a reason is found, so this patch also includes an alternation
between Top and Bot: if Bot was picked on the previous pickNode, then priority
is given to Top (in case no reason is found to prefer one or another, Top will
get picked). Vice-versa: if Top was picked last, then Bot will get prioritized.

As to not mess with the other targets, this extra behavior is only activated if
the target implements alternateTopBottomBidirectional(); we set it to true for
KVX.

Test Plan: CI + performance comparison

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D32597